More nuanced logging for response body copying errors
When Typhon encounters an error copying a response body, it's not helpful for an error that means "client disconnected so there's nowhere to send this" to be reported as an exceptional circumstance – the server is not misbehaving at all. These kinds of situations are now reported as debug-level events (it's still helpful for them to be reported if someone is actually looking for them).